requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
groq==0.9.0
pyyaml==6.0.1
pydantic==2.7.4
python-multipart==0.0.9
